# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "MarketPlace Standard App API",
    "description": "Standard app API for the MarketPlace applications.",
    "contact": {
      "name": "The Materials MarketPlace Consortium",
      "url": "https://www.materials-marketplace.eu/",
      "email": "dirk.helm@iwm.fraunhofer.de"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/data": {
      "get": {
        "tags": [
          "DataSource",
          "DataSink"
        ],
        "summary": "List Collections",
        "description": "List all collections.",
        "operationId": "listCollections",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Listcollections",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionListItemModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "204": {
            "description": "No collections found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      }
    },
    "/data/{collection_name}": {
      "get": {
        "tags": [
          "DataSource"
        ],
        "summary": "List Datasets",
        "description": "List all datasets.",
        "operationId": "listDatasets",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Listdatasets",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "204": {
            "description": "No datasets found."
          },
          "404": {
            "description": "Container not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "DataSink"
        ],
        "summary": "Create Or Update Collection",
        "description": "Create or update a collection.\n\nTo add custom metadata, add keys to the header of the form:\n\n- X-Object-Meta-name: value\n\nWhere 'name' is the name of the metadata key and 'value' is the\ncorresponding value.\n\nNote: This operation is in compliance with the OpenStack Swift object\nstorage API:\nhttps://docs.openstack.org/api-ref/object-store/index.html#create-container\n",
        "operationId": "createOrUpdateCollection",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Collection has been created."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "202": {
            "description": "Collection has been updated."
          },
          "400": {
            "description": "Bad request."
          },
          "507": {
            "description": "Insufficient storage."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "DataSink"
        ],
        "summary": "Delete Collection",
        "description": "Delete an empty collection.",
        "operationId": "deleteCollection",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Collection has been deleted."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "404": {
            "description": "Collection not found."
          },
          "409": {
            "description": "Collection is not empty."
          },
          "422": {
            "description": "Validation error."
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      },
      "head": {
        "tags": [
          "DataSource"
        ],
        "summary": "Get Collection Metadata",
        "description": "Get the metadata for a collection.",
        "operationId": "getCollectionMetadata",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Normal response."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "404": {
            "description": "Not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      }
    },
    "/data/": {
      "put": {
        "tags": [
          "DataSink"
        ],
        "summary": "Create Collection",
        "description": "Create a collection.\n\nTo add custom metadata, add keys to the header of the form:\n\n- X-Object-Meta-name: value\n\nWhere 'name' is the name of the metadata key and 'value' is the\ncorresponding value.\n\nNote: This operation is in compliance with the OpenStack Swift object\nstorage API:\nhttps://docs.openstack.org/api-ref/object-store/index.html#create-container\n",
        "operationId": "createCollection",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "query"
          }
        ],
        "responses": {
          "201": {
            "description": "Collection has been created."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "400": {
            "description": "Bad request."
          },
          "507": {
            "description": "Insufficient storage."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      }
    },
    "/data/{collection_name}/": {
      "put": {
        "tags": [
          "DataSink"
        ],
        "summary": "Create Dataset",
        "description": "Create a dataset.\n\nTo add custom metadata, add keys to the header of the form:\n\n- X-Object-Meta-name: value\n\nWhere 'name' is the name of the metadata key and 'value' is the\ncorresponding value.\n\nNote: This operation is in compliance with the OpenStack Swift object\nstorage API:\nhttps://docs.openstack.org/api-ref/object-store/index.html#create-or-replace-object\n",
        "operationId": "createDataset",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Dataset Name",
              "minLength": 1,
              "type": "string"
            },
            "name": "dataset_name",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createDataset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "507": {
            "description": "Insufficient storage."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "DataSink"
        ],
        "summary": "Create Dataset Metadata",
        "description": "Create or replace dataset metadata.\n\nNote: This operation is in compliance with the OpenStack Swift object\nstorage API:\nhttps://docs.openstack.org/api-ref/object-store/index.html#create-or-update-object-metadata",
        "operationId": "createDatasetMetadata",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Collection Name",
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "name": "collection_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Dataset Name",
              "minLength": 1,
              "type": "string"
            },
            "name": "dataset_name",
            "in": "query"
          }
        ],
        "responses": {
          "202": {
            "description": "Dataset metadata has been created."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "501": {
            "description": "Not implemented."
          },
          "503": {
            "description": "Service unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_createDataset": {
        "title": "Body_createDataset",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "CollectionListItemModel": {
        "title": "CollectionListItemModel",
        "required": [
          "count",
          "bytes",
          "name"
        ],
        "type": "object",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "last_modified": {
            "title": "Last Modified",
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DatasetCreateResponse": {
        "title": "DatasetCreateResponse",
        "required": [
          "last_modified"
        ],
        "type": "object",
        "properties": {
          "last_modified": {
            "title": "Last Modified",
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DatasetModel": {
        "title": "DatasetModel",
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "minLength": 1,
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "last_modified": {
            "title": "Last Modified",
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "AuthTokenBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}